Move Address to seedu.addressbook.data.address package
