+++ src/seedu/addressbook/data/person/Person.java
@@ -1,5 +1,6 @@
 package seedu.addressbook.data.person;
 
+import seedu.addressbook.data.address.Address;
 import seedu.addressbook.data.tag.UniqueTagList;
 
 import java.util.Objects;
