+++ src/seedu/addressbook/storage/jaxb/AdaptedPerson.java
@@ -1,6 +1,7 @@
 package seedu.addressbook.storage.jaxb;
 
 import seedu.addressbook.common.Utils;
+import seedu.addressbook.data.address.Address;
 import seedu.addressbook.data.exception.IllegalValueException;
 import seedu.addressbook.data.tag.Tag;
 import seedu.addressbook.data.tag.UniqueTagList;
